requirements.txt
streamlit
pandas
google-generativeai
openpyxl
jinja2
